Moved @tryghost/email-analytics-provider-mailgun to services folder
fix https://linear.app/ghost/issue/ENG-2092/email-analytics-provider-mailgun

- this commit moves the lib code and tests to the core folder so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
