+++ compose.yml
@@ -48,7 +48,6 @@ services:
       - node_modules_ghost_domain-events:/home/ghost/ghost/domain-events/node_modules:delegated
       - node_modules_ghost_donations:/home/ghost/ghost/donations/node_modules:delegated
       - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
-      - node_modules_ghost_email-analytics-provider-mailgun:/home/ghost/ghost/email-analytics-provider-mailgun/node_modules:delegated
       - node_modules_ghost_email-analytics-service:/home/ghost/ghost/email-analytics-service/node_modules:delegated
       - node_modules_ghost_email-content-generator:/home/ghost/ghost/email-content-generator/node_modules:delegated
       - node_modules_ghost_email-events:/home/ghost/ghost/email-events/node_modules:delegated
@@ -210,7 +209,6 @@ volumes:
   node_modules_ghost_donations: {}
   node_modules_ghost_email-addresses: {}
   node_modules_ghost_email-analytics-provider-mailgun: {}
-  node_modules_ghost_email-analytics-service: {}
   node_modules_ghost_email-content-generator: {}
   node_modules_ghost_email-events: {}
   node_modules_ghost_email-service: {}
