+++ ghost/core/core/server/services/email-analytics/EmailAnalyticsProviderMailgun.js
