--- ghost/email-analytics-provider-mailgun/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/EmailAnalyticsProviderMailgun');
