--- ghost/email-analytics-provider-mailgun/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@tryghost/email-analytics-provider-mailgun",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura --check-coverage --100 -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "mocha": "10.2.0",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/email-analytics-service": "0.0.0",
-    "@tryghost/mailgun-client": "0.0.0",
-    "moment": "2.29.1"
-  }
-}
